website/requirements.txt
django==1.11.8
celery==3.1.25
uwsgi==2.0.17
django-celery==3.2.2
redis==2.10.6
django-redis==4.9.0
asgi-redis==1.4.3
mysqlclient==1.3.12
paramiko==2.4.0
pexpect==4.5.0
